[Python]: fix inception

+++ python/flexflow_dataloader.cc
@@ -228,7 +228,7 @@ void ImgDataLoader4D::load_entire_dataset(const Task *task,
   FILE* file = fopen(alexnet->dataset_path.c_str(), "rb");
   unsigned char* buffer = (unsigned char*) malloc(3073);
   unsigned char* image = (unsigned char*) malloc(3 * height * width);
-  for (int i = 0; i < num_samples; i++) {
+  for (off_t i = 0; i < num_samples; i++) {
     size_t ret = fread(buffer, sizeof(unsigned char), 3073, file);
     assert(ret = 3073);
     if (i == 0) {
@@ -243,9 +243,9 @@ void ImgDataLoader4D::load_entire_dataset(const Task *task,
     label_ptr[i] = buffer[0];
     nearest_neigh(image, buffer + 1, height, width,
                   origHeight, origWidth, heightScale, widthScale);
-    int input_offset = i * 3 * height * width;
-    int image_offset = 0;
-    for (int h = 0; h < 3*height*width; h++)
+    off_t input_offset = i * 3 * height * width;
+    off_t image_offset = 0;
+    for (off_t h = 0; h < 3*height*width; h++)
         input_ptr[input_offset++] = static_cast<float>(image[image_offset++]) / 255;
   }
   printf("Finish loading %d samples from %s\n",
